Refactor WorldGuard integration to use WorldGuard 7 handler directly
